+++ server/tasks/sync/cdn.ts
@@ -13,14 +13,14 @@ export default defineTask({
       // update cover -> https://cdn.redcatpictures.com/media/w_1620&h_1080/product-photo-033-033
 
       // asset.properties.Type.select.name === 'Photo' && asset.properties.Status.status.name === 'Plan'
-      if (!(asset.properties.Status.status.name === 'Plan')) continue
+      if (asset.properties.Type.select.name === 'Photo') continue
 
       const slug = asset.properties.Slug?.formula?.string === 'featured-video-000-000' ? asset.properties.Slug?.formula?.string + '-landscape' : asset.properties.Slug?.formula?.string
       const [aW, aH] = asset.properties['Aspect ratio'].select.name.split(':').flatMap((item) => parseInt(item))
       const aspectRatio = aW / aH
       const { width: coverWidth, height: coverHeight } = calculateDimension(1080, aspectRatio)
 
-      let updateCoverURL = `${config.public.cdnUrl}/image/w_${coverWidth}&h_${coverHeight}/${slug}`
+      let updateCoverURL = `${config.public.cdnUrl}/image/s_${coverWidth}x${coverHeight}/${slug}`
 
       let coverExists = false
       try {
@@ -35,8 +35,32 @@ export default defineTask({
         continue
       }
 
-      updateCoverURL = `https://cdn.redcatpictures.com/media/image/w_${coverWidth}&h_${coverHeight}/${slug}`
+      updateCoverURL = `https://cdn.redcatpictures.com/media/image/s_${coverWidth}x${coverHeight}/${slug}`
       console.log('🍃 Updating', { slug, updateCoverURL })
+
+      const metaData =
+        import.meta.env.NODE_ENV === 'production'
+          ? await $fetch<{
+              format: {
+                filename: string
+                formatName: string
+                duration: number
+                size: number
+                bitRate: number
+              }
+              stream: {
+                codecName: string
+                codecType: string
+                width: number
+                height: number
+                bitRate: number
+                duration: number
+                frameRate: number
+              }
+            }>(`/media/videos/${slug}`, {
+              baseURL: 'https://cdn.redcatpictures.com/api',
+            })
+          : undefined
 
       await notion.pages.update({
         page_id: asset.id,
@@ -52,6 +76,11 @@ export default defineTask({
               name: asset.properties.Status.status.name, //coverExists ? (asset.properties.Status.status.name !== 'Plan' ? asset.properties.Status.status.name : 'Draft') : 'Plan',
             },
           },
+          ...(metaData && {
+            Additional: {
+              rich_text: [{ text: { content: `${JSON.stringify({ duration: metaData.format.duration })}` } }],
+            },
+          }),
         },
       })
     }